% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handler.R
\name{list_csv_files}
\alias{list_csv_files}
\title{List all CSV files in the data/ folder, which is where the activation data
will be}
\usage{
list_csv_files()
}
\value{
A list of all the file names
}
\description{
List all CSV files in the data/ folder, which is where the activation data
will be
}
