% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_activations.R
\name{plot_activations}
\alias{plot_activations}
\title{plot_activations}
\usage{
plot_activations(file_path)
}
\arguments{
\item{file_path}{Path to the CSV file.}
}
\value{
A ggplot.
}
\description{
these are the functions for plotting activations
}
