% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-activations.R
\name{plot_activations}
\alias{plot_activations}
\title{Plot the activations resulting from a given feature on the given amino acid, currently just 0 and 1.}
\usage{
plot_activations(file_path)
}
\arguments{
\item{file_path}{Path to the CSV file.}
}
\value{
A ggplot.
}
\description{
Plot the activations resulting from a given feature on the given amino acid, currently just 0 and 1.
}
